Convert SelectField to CSS-in-JS
